+++ TimberCommons/package/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.3 (10 Nov 2023)
+* [Fix #17] Support contaminated water (Update 5).
+* This version is not compatible with Update 4.
+
+# v1.2 (30 Sep 2023)
+* [Fix #15] Support new water simulation logic in Update 5.
+
 # v1.1 (27 Jul 2023)
 * [Fix #4] In rare cases the game crashes when valve gets built.
 * Add `deDE` localization from @juf0816.